requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
ollama>=0.2.0
pymupdf>=1.24.0
pypdf>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
python-slugify>=8.0.0
